chore: Add 3rd generation iPad pro to device->model map (#618)
## Description
I've added the 3rd generation Ipad pro to the list with available devices. This fixes a problem where it just returns "iPad" for these devices.

I've specifically added 3th generation to the name of the device. This way you can actually tell the difference between the multiple versions of the iPad Pro. This is a must have because it's the only way to tell the difference between the iPad pro with physical home button (generation 1 and 2) and the iPad pro without physical home button (generation 3). I kept the name of generation 2 the same for now but I can rename it for consistency.