ci(.github/workflows/release.yml): using `GH_TOKEN` as env and secret (#1187)
replaces `GITHUB_TOKEN` environment variable w/ `GH_TOKEN` and updates the value to use the secret
of the same name. `GITHUB_TOKEN` secret is a special secret that can't be set, so if you're writing
into a protected branch, you'll need to add a PAT of a user that can write to said protected
branches.