build: workaround obscure types issue for transitive relesae-it dep
tsc cannot find the types for parse-path, but they the are in the
package itself. stackoverflow recommends just overriding typeRoots
and moving on with life

if you do that, then ts-jest blows up though, so we need a separate config
for building (bob, tsc) and for jest.

Finally, jest doesn't like to find custom tsconfigs easily, but tsc and bob do,
so:

1- default tsconfig has the config needed for jest - which is the original one
2- new "build"-only tsconfig just extends the original one with typeroots
3- tsc and bob are pointed to the new one

No, I do not like any of this, but...it works